src/app/services/user.service.ts
// src/app/services/user.service.ts

import { Injectable } from '@angular/core';
import { HttpClient } from '@angular/common/http';
import { Observable } from 'rxjs';
import { User } from '../models/user.model';

@Injectable({
  providedIn: 'root',
})
export class UserService {
  private apiUrl = 'https://randomuser.me/api/?seed=nuvalence';

  constructor(private http: HttpClient) {}

  fetchUsers(page: number): Observable<any> {
    const url = `${this.apiUrl}&page=${page}&results=10`;
    return this.http.get<any>(url);
  }

  setUsers(users: User[]): void {
    localStorage.setItem('users', JSON.stringify(users));
  }

  getUsers(): User[] {
    const users = localStorage.getItem('users');
    return users ? JSON.parse(users) : [];
  }

  getUserById(id: string): User | undefined {
    const users = this.getUsers();
    return users.find(user => user.login.uuid === id);
  }
}